Factor out createContext from createSystem
